edge/requirements.txt
firebase-admin==6.6.0
python-dotenv==1.0.1
RPi.GPIO==0.7.1; platform_system == "Linux"
